# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AcuFi Backend API",
    "description": "AcuFi Backend API with professional architecture",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/document-types/": {
      "get": {
        "tags": [
          "document-types"
        ],
        "summary": "Get Document Types",
        "description": "Retrieve active document types.\nReturns the Id, DocumentType and Description for each record.",
        "operationId": "get_document_types_api_v1_document_types__get",
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Skip"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentTypePublic"
                  },
                  "title": "Response Get Document Types Api V1 Document Types  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/document-types/{document_type_id}": {
      "get": {
        "tags": [
          "document-types"
        ],
        "summary": "Get Document Type",
        "description": "Get document type by ID.",
        "operationId": "get_document_type_api_v1_document_types__document_type_id__get",
        "parameters": [
          {
            "name": "document_type_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Document Type Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentType"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/document-types/by-name/{document_type_name}": {
      "get": {
        "tags": [
          "document-types"
        ],
        "summary": "Get Document Type By Name",
        "description": "Get document type by name.",
        "operationId": "get_document_type_by_name_api_v1_document_types_by_name__document_type_name__get",
        "parameters": [
          {
            "name": "document_type_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Type Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentType"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/document-detail": {
      "get": {
        "tags": [
          "document-detail"
        ],
        "summary": "Fetch Document Details",
        "operationId": "fetch_document_details_api_v1_document_detail_get",
        "parameters": [
          {
            "name": "document_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "maximum": 4,
              "minimum": 1,
              "title": "Document Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDetailRead"
                  },
                  "title": "Response Fetch Document Details Api V1 Document Detail Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "document-detail"
        ],
        "summary": "Update Is Critical",
        "operationId": "update_is_critical_api_v1_document_detail_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentDetailUpdateCritical"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DocumentDetailRead": {
        "properties": {
          "Id": {
            "type": "integer",
            "title": "Id"
          },
          "DocId": {
            "type": "integer",
            "title": "Docid"
          },
          "FieldKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fieldkey"
          },
          "FieldLabelToDisplay": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fieldlabeltodisplay"
          },
          "isCritical": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iscritical"
          }
        },
        "type": "object",
        "required": [
          "Id",
          "DocId"
        ],
        "title": "DocumentDetailRead"
      },
      "DocumentDetailUpdateCritical": {
        "properties": {
          "Id": {
            "type": "integer",
            "title": "Id"
          },
          "isCritical": {
            "type": "boolean",
            "title": "Iscritical"
          }
        },
        "type": "object",
        "required": [
          "Id",
          "isCritical"
        ],
        "title": "DocumentDetailUpdateCritical"
      },
      "DocumentType": {
        "properties": {
          "DocumentType": {
            "type": "string",
            "title": "Documenttype"
          },
          "Description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "ExpectedJson": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expectedjson"
          },
          "CreatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdby"
          },
          "LastModifiedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastmodifiedby"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "default": true
          },
          "VerticalThresholdScore": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Verticalthresholdscore"
          },
          "Id": {
            "type": "integer",
            "title": "Id"
          },
          "CreatedOn": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdon"
          },
          "LastModifiedOn": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastmodifiedon"
          }
        },
        "type": "object",
        "required": [
          "DocumentType",
          "Id"
        ],
        "title": "DocumentType"
      },
      "DocumentTypePublic": {
        "properties": {
          "Id": {
            "type": "integer",
            "title": "Id"
          },
          "DocumentType": {
            "type": "string",
            "title": "Documenttype"
          },
          "Description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "Id",
          "DocumentType"
        ],
        "title": "DocumentTypePublic",
        "description": "Public-facing schema that exposes Id, DocumentType and Description"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}